Tidy the address book after the owner rewrite
Five things the rewrite left behind.

Spree::Address still carried its own owner-type chain for reading a
default slot, so the concern that exists to declare those columns once
declared them twice — and a third owner would have had to be registered
in both, with only one failing loudly.

The admin controller still saved the row itself and poked the pointers
afterwards, while the store half had moved to the address services. Two
write paths for one resource, and only one of them got the dedup and
replacement handling. Both go through the services now, which leaves the
two controller concerns holding the same three methods, so they collapse
into one.

The inherited reading list becomes Company#address_book. The scope was
the only place in the codebase that widens for index rather than
narrowing, and the security property rested on a comment explaining it;
naming the relation puts the asymmetry in the method names instead.

A promotion writes its own in-memory attributes, so only a release needs
the reload. And the three-state flag rule is proved once against both
owners in the concern spec rather than three times through HTTP.
