+++ spree/api/app/controllers/spree/api/v3/store/companies/addresses_controller.rb
@@ -7,7 +7,7 @@ module Spree
           # reached through its node, so standing over that node is the only
           # thing that authorizes reading or writing one.
           class AddressesController < BaseController
-            include Spree::Api::V3::Store::Concerns::CompanyAddressDefaults
+            include Spree::Api::V3::CompanyAddressWrites
 
             # POST /api/v3/store/companies/:company_id/addresses
             def create
@@ -57,22 +57,13 @@ module Spree
               Spree.api.address_serializer
             end
 
-            # Reading a book and keeping one are different rights.
-            #
-            # Listing shows the node's own sites plus the ones it inherits —
-            # the same self-and-ancestors chain its default address is
-            # prefilled from, and the same one checkout accepts an id from, so
-            # a division ships to its headquarters' addresses.
-            #
-            # Writing stays with the node that owns the row. Standing over a
-            # division is not standing over its parent, and the Store API
-            # authorizes purely by what this scope returns — so widening it for
-            # every action would let a division member rename, delete, or
-            # re-point the defaults of the company above them.
+            # Reading a book and keeping one are different rights: the listing
+            # includes what this node inherits, writes stay with the node that
+            # owns the row. The Store API authorizes purely by what this scope
+            # returns, so widening it for every action would let a division
+            # member rename or re-point its parent's entries.
             def scope
-              return @parent.addresses unless action_name == 'index'
-
-              Spree::Address.where(owner_type: 'Spree::Company', owner_id: @parent.self_and_ancestors.map(&:id))
+              action_name == 'index' ? @parent.address_book : @parent.addresses
             end
 
             def parent_association
