+++ spree/api/spec/controllers/spree/api/v3/admin/companies/addresses_controller_spec.rb
@@ -88,8 +88,8 @@ RSpec.describe Spree::Api::V3::Admin::Companies::AddressesController, type: :con
       expect(company.reload.default_bill_address_id).to eq(entry.id)
     end
 
-    # Clearing only unsets the pointer when this row is the one holding it —
-    # another entry's default is not this request's business.
+    # The three-state rule itself is proved once, against both owner types,
+    # in the HasAddressBook spec; this only shows the flag reaches the model.
     it 'clears the default it holds' do
       company.update!(default_bill_address: entry)
 
@@ -98,17 +98,6 @@ RSpec.describe Spree::Api::V3::Admin::Companies::AddressesController, type: :con
 
       expect(response).to have_http_status(:ok)
       expect(company.reload.default_bill_address_id).to be_nil
-    end
-
-    it 'leaves another entry default alone when cleared' do
-      other = create(:company_address, owner: company, label: 'Dock')
-      company.update!(default_bill_address: other)
-
-      patch :update, params: { company_id: company.prefixed_id, id: entry.prefixed_id, default_billing: false },
-                     as: :json
-
-      expect(response).to have_http_status(:ok)
-      expect(company.reload.default_bill_address_id).to eq(other.id)
     end
   end
 end
