+++ spree/core/app/models/spree/company.rb
@@ -188,6 +188,19 @@ module Spree
       default_address(:default_ship_address)
     end
 
+    # Everything this node may ship to: its own sites and the ones it inherits
+    # from the nodes above it — the same chain its default address is
+    # prefilled from, and the one checkout accepts an address id from.
+    #
+    # Deliberately wider than {#addresses}: a division reads its headquarters'
+    # entries but keeps none of them, so this is a reading list and never a
+    # write target.
+    #
+    # @return [ActiveRecord::Relation]
+    def address_book
+      Spree::Address.where(owner_type: 'Spree::Company', owner_id: self_and_ancestors.map(&:id))
+    end
+
     private
 
     def ensure_can_be_deleted
